split github CI workflow
This commit splits the github CI workflow into two parts. The
`validation` portion handles linting, testing, license header checks,
etc. The `pr-comment` portion updates the PR with the results of the
`validation` portion.

This workflow split allows for incoming PRs with workflow changes to use
said changes as part of the their validation workflow. Previously, the
workflows operated in the `pull_request_target` context, instead of the
`pull_request` context in order to allow the ability to write comments
to the PR, but had the effect of using the workflows from the target
branch to validate the incoming PR.

Signed-off-by: Rob Nester <rnester@redhat.com>