fix skipping VSA generation in demo
This commit fixes an issue in the previous version of the demo where the
VSA generation would be skipped as an existing VSA attestation had been
generated in a previous execution of the demo. This commit has
refactored the demo code to include a build time argument (`BUILD_ID`)
that ensures that each execution of the demo generates a unique version
of the image, ensuring that VSA generation is performed.

Signed-off-by: Rob Nester <rnester@redhat.com>